refatorando style

+++ places-service/app/main.py
@@ -60,7 +60,7 @@ def get_places(
     category: str | None = None,
     occasion: str | None = None,
     price_level: int | None = None,
-    min_rating: int | None = None,
+    min_rating: float | None = None,
     event_type: str | None = None,
     session: Session = Depends(get_session)
 ):
